ACC| refactor code

+++ app/src/main/java/com/pokotechnologies/beeceptor/app/model/Main.kt
@@ -6,10 +6,10 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class Main(
-    @SerialName(value = "temp") val temp: Int,
+    @SerialName(value = "temp") val temp: Double,
     @SerialName(value = "feels_like") val feels_like: Double,
-    @SerialName(value = "temp_min") val temp_min: Int,
-    @SerialName(value = "temp_max") val temp_max: Int,
+    @SerialName(value = "temp_min") val temp_min: Double,
+    @SerialName(value = "temp_max") val temp_max: Double,
     @SerialName(value = "pressure") val pressure: Int,
     @SerialName(value = "humidity") val humidity: Int
 )
